added scheduler settings (#10)

+++ src/strategies/random.cc
@@ -22,9 +22,9 @@ namespace coyote
 
 	size_t Random::next()
 	{
-		const uint64_t x = state_x;
-		uint64_t y = state_y;
-		const uint64_t result = state_x + y;
+		const size_t x = state_x;
+		size_t y = state_y;
+		const size_t result = state_x + y;
 
 		y ^= x;
 		state_x = rotl(x, 24) ^ y ^ (y << 16);
